+++ src/reshuffle/reshuffle.h
@@ -29,7 +29,7 @@ public:
 	void write_data_chi(ifstream&,ofstream&);
 	void write_slim_data(ifstream&,ofstream&);
 	void write_herest(ifstream&,ofstream&);
-	int64_t herest_startpos;
+	long long herest_startpos;
 	int est_shift(int);
 	int est_beta_shift(int);
 	void run(Parameters);
